requirements.txt
Django==3.0.3
asgiref==3.2.3
django-js-asset==1.2.2
django-mptt==0.11.0
django-taggit==1.2.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
